+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * properly add `vertebrae_body` model
 * add `--roi_subset_robust` argument
 * add `--fastest` argument
-* allow `mps` as device
+* allow `mps` as device (but not working yet)
 
 
 ## Release 2.0.5