+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Master
+
+
+## Release 2.1.0
 * Bugfix: add flush to DummyFile
 * Require python >= 3.9 in setup.py
 * properly add `vertebrae_body` model
@@ -10,8 +13,8 @@
 * from python api also return nifti with label map in header
 * input to python api can be a Nifti1Image object or a file path
 * upgrade to `nnunetv2>=2.2.1`
+* for `total` task use nnU-Net `step_size=0.8` instead of `0.5` for faster runtime while only decreasing dice by 0.001
 * minor edits and bugfixes
-* use nnU-Net `step_size=0.8` instead of `0.5` for faster runtime while only decreasing dice by 0.001
 
 
 ## Release 2.0.5